refactor(mariadb): rename file previously with mysql in name
